backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic[email]==2.9.2
python-dotenv==1.0.1
transformers==4.44.2
torch==2.4.1
langchain==0.3.3
langchain-core==0.3.10
langchain-community==0.3.2
langchain-groq==0.2.0
groq==0.4.2
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
httpx==0.24.1
